+++ code/vertexEdge.h
@@ -16,7 +16,7 @@ using namespace std;
 
 class Edge;
 
-#define INF std::numeric_limits<double>::max()
+#define INF std::numeric_limits<int>::max()
 
 class Vertex {
 public:
@@ -28,7 +28,7 @@ public:
     bool isVisited() const;
     bool isProcessing() const;
     unsigned int getIndegree() const;
-    double getDist() const;
+    int getDist() const;
     Edge *getPath() const;
     std::vector<Edge *> getIncoming() const;
     Station getStation() const;
@@ -39,7 +39,7 @@ public:
     void setVisited(bool visited);
     void setProcesssing(bool processing);
     void setIndegree(unsigned int indegree);
-    void setDist(double dist);
+    void setDist(int dist);
     void setPath(Edge *path);
     Edge * addEdge(Vertex *dest, int capacity, string service);
     bool removeEdge(int destID);
@@ -55,7 +55,7 @@ private:
     bool visited = false; // used by DFS, BFS, Prim ...
     bool processing = false; // used by isDAG (in addition to the visited attribute)
     unsigned int indegree; // used by topsort
-    double dist = 0;
+    int dist = 0;
     Edge *path = nullptr;
 
     std::vector<Edge *> incoming; // incoming edges
@@ -77,11 +77,11 @@ public:
     bool isSelected() const;
     Vertex * getOrig() const;
     Edge *getReverse() const;
-    double getFlow() const;
+    int getFlow() const;
 
     void setSelected(bool selected);
     void setReverse(Edge *reverse);
-    void setFlow(double flow);
+    void setFlow(int flow);
 private:
     Vertex * dest; // destination vertex
     int capacity; // edge weight, can also be used for capacity
@@ -93,7 +93,7 @@ private:
     Vertex *orig;
     Edge *reverse = nullptr;
 
-    double flow; // for flow-related problems
+    int flow; // for flow-related problems
 };
 
 #endif //DA_VERTEXEDGE_H
